Moving material-tooltip folder up to ./components
